# Reduce runtime for and allocations from BaseConsoleLogger.IndentString