--- a/src/Build/Logging/OptimizedStringIndenter.cs
+++ b/src/Build/Logging/OptimizedStringIndenter.cs
@@ -3,8 +3,12 @@
 
 using System;
 using System.Buffers;
+
 #if NET7_0_OR_GREATER
 using System.Runtime.CompilerServices;
+#else
+using System.Text;
+using Microsoft.Build.Framework;
 #endif
 
 namespace Microsoft.Build.BackEnd.Logging;