… src/Build/Logging/ParallelLogger/ConsoleOutputAligner.cs
… src/Build/Logging/ParallelLogger/ParallelConsoleLogger.cs
… src/Build/Microsoft.Build.csproj
… src/Framework/IStringBuilderProvider.cs