Fixes [Bug 1828095](https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1828095): [GCPauseWatson] MSBuild: BaseConsoleLogger.IndentString uses String.Split and StringBuilder to create an indented string

GCPauseWatson telemetry showed that `BaseConsoleLogger.IndentString` is responsible for allocating 1.99 GB of memory in VS at the 95th percentile.  This change aims to address that by replacing the multiple allocations of `String.Split` and `StringBuilder` with code that makes a single allocation of a right-sized string and then builds the indented string directly in the new string's buffer.

Benchmarking shows that -- depending on the target framework -- the new algorithm runs in 50-75% of the time of the code it replaces, and allocates 15-30% of the memory.  The optimizations have more of an impact on net472 than on net7.0 due to the `String` and `StringBuilder` optimizations included in net7.0.

<details>
```
BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22631.1900)
Intel Core i9-10900 CPU 2.80GHz, 1 CPU, 20 logical and 10 physical cores
  [Host]               : .NET Framework 4.8.1 (4.8.9166.0), X64 RyuJIT VectorSize=256
  .NET 7.0             : .NET 7.0.7 (7.0.723.27404), X64 RyuJIT AVX2
  .NET Framework 4.7.2 : .NET Framework 4.8.1 (4.8.9166.0), X64 RyuJIT VectorSize=256


|    Method |              Runtime | StringCount |         Mean | Ratio |     Gen0 |   Gen1 | Allocated | Alloc Ratio |
|----------:|---------------------:|------------:|-------------:|------:|---------:|-------:|----------:|------------:|
|  Original |             .NET 7.0 |           2 |   1,165.9 ns |  1.00 |   0.3052 |      - |    3192 B |        1.00 |
| Optimized |             .NET 7.0 |           2 |     803.1 ns |  0.69 |   0.0925 |      - |     976 B |        0.31 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |           2 |   1,832.6 ns |  1.00 |   1.0567 | 0.0038 |    6660 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |           2 |     909.1 ns |  0.50 |   0.1554 |      - |     979 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |           4 |   1,857.7 ns |  1.00 |   0.5150 |      - |    5400 B |        1.00 |
| Optimized |             .NET 7.0 |           4 |   1,383.5 ns |  0.74 |   0.1545 |      - |    1632 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |           4 |   3,119.1 ns |  1.00 |   1.7738 | 0.0038 |   11177 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |           4 |   1,603.1 ns |  0.51 |   0.2613 |      - |    1653 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |           8 |   3,642.3 ns |  1.00 |   0.9766 |      - |   10232 B |        1.00 |
| Optimized |             .NET 7.0 |           8 |   2,716.4 ns |  0.75 |   0.2937 |      - |    3096 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |           8 |   5,984.5 ns |  1.00 |   3.3417 | 0.0076 |   21070 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |           8 |   3,130.6 ns |  0.52 |   0.4997 |      - |    3145 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |          16 |   6,457.6 ns |  1.00 |   1.7242 |      - |   18096 B |        1.00 |
| Optimized |             .NET 7.0 |          16 |   4,698.2 ns |  0.73 |   0.5188 |      - |    5424 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |          16 |  10,747.7 ns |  1.00 |   5.8136 | 0.0153 |   36628 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |          16 |   5,735.8 ns |  0.53 |   0.8698 |      - |    5504 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |          32 |  12,028.4 ns |  1.00 |   3.1891 |      - |   33360 B |        1.00 |
| Optimized |             .NET 7.0 |          32 |   9,267.4 ns |  0.77 |   0.9460 |      - |    9920 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |          32 |  20,126.2 ns |  1.00 |  10.6201 | 0.0305 |   66918 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |          32 |  10,968.9 ns |  0.54 |   1.6022 |      - |   10102 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |          64 |  24,792.3 ns |  1.00 |   6.3477 |      - |   66552 B |        1.00 |
| Optimized |             .NET 7.0 |          64 |  18,397.9 ns |  0.74 |   1.8921 |      - |   19832 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |          64 |  40,532.2 ns |  1.00 |  21.1792 |      - |  133610 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |          64 |  22,522.4 ns |  0.56 |   3.1738 |      - |   20131 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |         128 |  50,012.8 ns |  1.00 |  12.2681 |      - |  128672 B |        1.00 |
| Optimized |             .NET 7.0 |         128 |  37,066.0 ns |  0.74 |   3.6011 |      - |   38136 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |         128 |  78,969.9 ns |  1.00 |  40.7715 |      - |  256999 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |         128 |  44,377.9 ns |  0.56 |   6.1646 |      - |   38827 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |         256 |  95,493.7 ns |  1.00 |  23.6816 |      - |  248544 B |        1.00 |
| Optimized |             .NET 7.0 |         256 |  71,305.1 ns |  0.75 |   6.9580 |      - |   73360 B |        0.30 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |         256 | 151,293.4 ns |  1.00 |  78.1250 |      - |  492317 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |         256 |  86,210.2 ns |  0.57 |  11.8408 |      - |   74645 B |        0.15 |
|           |                      |             |              |       |          |        |           |             |
|  Original |             .NET 7.0 |         512 | 190,107.9 ns |  1.00 |  47.1191 |      - |  494032 B |        1.00 |
| Optimized |             .NET 7.0 |         512 | 142,247.7 ns |  0.75 |  13.9160 |      - |  145728 B |        0.29 |
|           |                      |             |              |       |          |        |           |             |
|  Original | .NET Framework 4.7.2 |         512 | 301,655.3 ns |  1.00 | 155.2734 |      - |  978898 B |        1.00 |
| Optimized | .NET Framework 4.7.2 |         512 | 177,835.2 ns |  0.59 |  23.4375 |      - |  148296 B |        0.15 |
```
</details>