@danmoseley 2023-06-22
@stephentoub  do we officially support working directly into a string like this (it is newly created)? I know we do it in StringBuilder but that's corelib.
Also do you see any opportunities to make this change more efficient still in this PR?