@stephentoub 2023-06-22
> do we officially support writing directly into a string like this (it is newly created)?

Most definitely no.