@danmoseley 2023-06-22
I guess then it should use StringBuilder and then ToString(), which is allowed to do it😊
And string.Create if multi targeting