@danmoseley 2023-06-22
> Zero guarantees about it working tomorrow.

Just curious, how could it break? Assuming it's a fresh string, not interned, no other references, it's fixed and they stay inside it.
Ah - because a future implementation could allocate on a ro page?