@stephentoub 2023-06-22
> Ah - because a future implementation could allocate on a ro page?

For example, yes. Or it could intern it and hand the same instance out to someone else who created a string of the same length and contents as provided to the ctor.  Or other such things.