@davkean 2023-06-22
> Most definitely no. It will work today. Zero guarantees about it working tomorrow.

There is no way this can be broken on .NET Framework given how many places this is done. This would break a lot. All usage I've seen over the years is via string that the person just created.