@danmoseley 2023-06-22
Is a pooled StringBuilder less desirable?

Either way there's no reason to do it on .NET Core.