@davkean 2023-06-22
Agreed on .NET Core, I personally missed that these APIs were added.

A pooled string builder could be a replacement, but not always convenient in larger strings where you don't want to put the resulting builder back into the pool to avoid leaking, it also chunks in 8K blocks for reasons I don't understand.