@rainersigwald 2023-06-22
Can you share your benchmark code please?