@danmoseley 2023-06-22
> It also chunks in 8K blocks for reasons I don't understand.

To avoid copies on resize and LOH I guess. Something we might consider in future is adding a mechanism to give a SB its buffers so we can pool nicely sized ones.