@davkean 2023-06-22
It is sized to be under LOH, but LOH is 85K. :) 

Anyway, if StringBuilder pooling was all that was needed, `String.Create` wouldn't be needed.