@danmoseley 2023-06-22
Pooling has a cost, and also assumes it will be used again. String create is just straight line cost.