@davkean 2023-06-22
Sure, I've added a link to the telemetry data in the bug, at its worse this path allocates 358 MB/s over a 60 second period, 70% of which ends up on the LOH. Anything is better than String.Split.