@ryanmolden 2023-06-22
> For this case, I'd really like to see benchmarks on the much more maintainable approach of avoiding `String.Split` in favor of using `StringBuilder.Append(fullstring, startIndex, count)` with a pooled `StringBuilder`.

StringBuilder is kind of terrible for these kinds of situations as it copies the incoming strings into char[]'s just to then allocate a .NET string off them, which means memory will be 2x the incoming string length (actually likely more due to how StringBuilder works internally).

When you are processing an incoming string and returning the result within the same call, in hot path areas (as GCPauseWatson has shown this to be), we are well past the point where we can choose the easiest solution, as we have ample evidence (i.e. thousands of results in GCPauseWatson of users experiencing long GC pauses due to heavy allocation activity in short time spans) showing it simply doesn't scale.

Simply eliminating String.Split is a definite win either way, but creating unnecessary char[]s that cause GC pressure and more frequent long duration GCs is not a winning strategy.