@danmoseley 2023-06-23
@stephentoub i wonder what we could do (in .NET Core only, of course) to help make it possible to get closer to the unsafe/unsupported performance seen here.