@davkean 2023-06-23
Does Benchmark.net show LOH? There were traces with LOH contribution here, so strings splits were greater than 42KB that you might want to check.